Modified reservation service project name
