+++ reservation-service/src/main/java/com/epam/reservation/repository/ReservationRepository.java
@@ -1,0 +1,5 @@
+package com.epam.reservation.repository;
+
+public interface ReservationRepository {
+
+}
