+++ reservation-service/src/main/java/com/epam/reservation/service/ReservationService.java
@@ -1,0 +1,5 @@
+package com.epam.reservation.service;
+
+public class ReservationService {
+
+}
