--- reservationservice/src/main/java/com/epam/reservation/repository/ReservationRepository.java
@@ -1,5 +1,0 @@
-package com.epam.reservation.repository;
-
-public interface ReservationRepository {
-
-}
