--- reservationservice/src/main/java/com/epam/reservation/service/ReservationService.java
@@ -1,5 +1,0 @@
-package com.epam.reservation.service;
-
-public class ReservationService {
-
-}
