--- reservationservice/src/main/resources/application.properties
@@ -1,18 +1,0 @@
-
-spring.application.name=reservation-service
-server.port=8084
-
-# data source
-spring.datasource.url= jdbc:mysql://localhost:3306/reservation
-spring.datasource.username=root
-spring.datasource.password=root
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-
-# JPA Releated
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL8Dialect
-spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
-
-#Schema will be created using schema.sql and data.sql files
-spring.jpa.hibernate.ddl-auto=update
